# openapi_for_chatgpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Transcript API",
    "version": "1.0.0",
    "description": "API to fetch YouTube video transcripts, extract video IDs, and provide a health check."
  },
  "servers": [
    {
      "url": "https://YOUR_PUBLIC_BACKEND_URL",
      "description": "Production server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Health check endpoint",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/extract-video-id": {
      "get": {
        "operationId": "extractVideoId",
        "summary": "Extract a YouTube video ID from a URL",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Full YouTube video URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful extraction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "videoId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL supplied"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/transcript": {
      "get": {
        "operationId": "getTranscript",
        "summary": "Retrieve the transcript for a YouTube video",
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "YouTube video identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "videoId": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "language": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    }
                  },
                  "required": ["videoId", "text", "source"]
        }
      }
    },
    "/transcript-from-url": {
      "get": {
        "operationId": "getTranscriptFromUrl",
        "summary": "Retrieve the transcript for a YouTube URL",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Full YouTube video URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "videoId": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "language": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    }
                  },
                  "required": ["videoId", "text", "source"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL supplied"
          },
          "404": {
            "description": "Transcript not available for this video"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
          "400": {
            "description": "Invalid YouTube video ID"
          },
          "404": {
            "description": "Transcript not available for this video"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}